feat(#61): represent the weapon-enchantment damage term
`Player::UpdateDamageDoneMods` (`Player.cpp:4991-5015`) adds
`ITEM_ENCHANTMENT_TYPE_DAMAGE` (2) `EffectScalingPoints` for the attack's
enchanted weapon, plus the same scaled by the weapon delay for
`ITEM_ENCHANTMENT_TYPE_TOTEM` (6) on shamans. The Rust
`SpellItemEnchantmentEntry` never loaded `EffectScalingPoints`, so enchanted
weapon damage was missing from the represented ranges.

- `SpellItemEnchantmentEntry` gains `effect_scaling_points: [f32; 3]`, read from
  DB2 field 3 (`DB2Structure.h:3573`); `Eq` is no longer derived because the
  field is `f32`. The twenty literals across the data/world fixtures gained the
  field explicitly.
- `represented_weapon_enchant_damage_like_cpp` resolves the attack's usable
  weapon, walks every enchantment slot of the item, and sums the damage/totem
  points; `represented_weapon_damage_flat_like_cpp` now returns `[f32; 3]` with
  the aura sum plus that term.
- `PlayerSpellBonusInputLikeCpp.weapon_damage_flat`,
  `PlayerStatSystemProjectionLikeCpp` and
  `PlayerEffectiveCombatStatsLikeCpp` move to `f32`, matching the C++
  `SetStatFlatModifier` float storage.

Evidence: the new session test runs the same enchanted sword fixture for a
shaman (class 7, permanent enchant with `effect = [DAMAGE, TOTEM]` and
`effect_scaling_points = [25.0, 10.0]`, 2.0 s weapon delay) and a warrior
(class 1): `weapon_damage_flat == [45.0, 0.0, 0.0]` and `[25.0, 0.0, 0.0]`
respectively; the existing wow-data flat/percentage test moves to the `f32`
values; wow-data --lib 752/0, wow-world --lib 3930/0/1; format,
`git diff --check`, the physical ratchet and `validation-v2 quick` (manifest
20260917T051430.821411Z-2355689-quick.json) pass. No live DB/restart/relogin QA.

+++ crates/wow-data/src/player/stats.rs
@@ -124,10 +124,10 @@ pub struct PlayerSpellBonusInputLikeCpp {
     /// base factor (mainhand/ranged 1.0, offhand 0.5) times the
     /// `SPELL_AURA_MOD_DAMAGE_PERCENT_DONE` (79) physical multiplier.
     pub weapon_damage_pct: [f32; 3],
-    /// `Unit::UpdateDamageDoneMods` (`Unit.cpp:8997-9027`) per attack: the
-    /// `SPELL_AURA_MOD_DAMAGE_DONE` (13) physical sum, published as the
-    /// `UNIT_MOD_DAMAGE_*` `TOTAL_VALUE`.
-    pub weapon_damage_flat: [i32; 3],
+    /// `Player::UpdateDamageDoneMods` (`Player.cpp:4965-5015`) per attack: the
+    /// `SPELL_AURA_MOD_DAMAGE_DONE` (13) physical sum plus the weapon
+    /// enchantment damage, published as the `UNIT_MOD_DAMAGE_*` `TOTAL_VALUE`.
+    pub weapon_damage_flat: [f32; 3],
     /// `SPELL_AURA_MOD_TARGET_RESISTANCE` (123) sums covering the full
     /// `SPELL_SCHOOL_MASK_SPELL`, published as `ModTargetResistance`.
     pub target_resistance_aura: i32,
@@ -153,7 +153,7 @@ impl Default for PlayerSpellBonusInputLikeCpp {
             damage_done_percent: [1.0; 7],
             healing_done_percent: 1.0,
             weapon_damage_pct: [1.0, 0.5, 1.0],
-            weapon_damage_flat: [0; 3],
+            weapon_damage_flat: [0.0; 3],
             versatility_bonus_aura: 0,
             target_resistance_aura: 0,
             item_spell_penetration: 0,
@@ -302,7 +302,7 @@ pub struct PlayerStatSystemProjectionLikeCpp {
     /// C++ `UNIT_MOD_DAMAGE_*` `TOTAL_PCT` per attack.
     pub weapon_damage_pct: [f32; 3],
     /// C++ `UNIT_MOD_DAMAGE_*` `TOTAL_VALUE` per attack.
-    pub weapon_damage_flat: [i32; 3],
+    pub weapon_damage_flat: [f32; 3],
 }
 
 /// C++ `Unit::CalculateMinMaxDamage` for the represented player weapon
@@ -358,7 +358,7 @@ pub fn effective_weapon_damage_ranges_like_cpp(
             .weapon_damage_flat
             .get(index)
             .copied()
-            .unwrap_or(0) as f32;
+            .unwrap_or(0.0);
         [
             ((weapon_min + ap_component + total_value) * total_pct).max(1.0),
             ((weapon_max + ap_component + total_value) * total_pct).max(1.0),
@@ -1503,7 +1503,7 @@ mod tests {
             attack_power_override_by_spell_power_pct: None,
             spell_bonus: PlayerSpellBonusInputLikeCpp {
                 weapon_damage_pct: [1.5, 0.75, 1.0],
-                weapon_damage_flat: [20, 40, 0],
+                weapon_damage_flat: [20.0, 40.0, 0.0],
                 ..Default::default()
             },
             rating_bonuses: [0.0; 32],
@@ -1577,7 +1577,7 @@ mod tests {
                 healing_done_percent: 2.0,
                 // Mainhand/ranged 1.0 and offhand 0.5 with a +50% physical aura.
                 weapon_damage_pct: [1.5, 0.75, 1.5],
-                weapon_damage_flat: [20, 20, 0],
+                weapon_damage_flat: [20.0, 20.0, 0.0],
                 versatility_bonus_aura: 200,
                 // `ModTargetResistance = aura - item penetration`.
                 target_resistance_aura: 20,
@@ -1608,7 +1608,7 @@ mod tests {
         assert_eq!(projection.override_ap_by_spell_power_percent, 0.0);
         assert_eq!(projection.versatility_bonus, 200.0);
         assert_eq!(projection.weapon_damage_pct, [1.5, 0.75, 1.5]);
-        assert_eq!(projection.weapon_damage_flat, [20, 20, 0]);
+        assert_eq!(projection.weapon_damage_flat, [20.0, 20.0, 0.0]);
 
         // `SPELL_AURA_OVERRIDE_SPELL_POWER_BY_AP_PCT` replaces both bonuses with
         // `int32(CalculatePct(GetTotalAttackPowerValue(BASE_ATTACK), pct) + 0.5)`:
